pyproject.toml
[project]
name = "becend"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi (>=0.126.0,<0.127.0)",
    "uvicorn (>=0.38.0,<0.39.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "alembic (>=1.17.2,<2.0.0)",
    "psycopg (>=3.3.2,<4.0.0)",
    "asyncio (>=4.0.0,<5.0.0)",
    "ollama (>=0.6.1,<0.7.0)"
]

[tool.poetry.group.test.dependencies]
flake8 = "^7.3.0"
flake8-pyproject = "^1.2.4"
pep8-naming = "^0.15.1"


[tool.flake8]
max-line-length = 120
ignore = [
    # W503: line break before binary operator
    "W503"
]
exclude = [
    ".git",
    "__pycache__",
    "conf",
    "venv",
    ".venv",
    ".vscode",
    ".idea",
]
per-file-ignores = [
    "migrations/env.py:F401,E402",
    "migrations/versions/*:W291,E122,E128,F401",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
